# openapi/openapi0008.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DripRate Server API",
    "description": "API for DripRate Server application",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:13000/"
    }
  ],
  "paths": {
...
    "/api/Collections/user/{userId}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "operationId": "Collections_GetUserCollections",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Collections_GetUserCollections_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/Feed/user/{userId}": {
      "get": {
        "tags": [
          "Feed"
        ],
        "operationId": "Feed_GetUserFeed",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Feed_GetUserFeed_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Feed_GetUserFeed_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Feed_GetUserFeed_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Feed_GetUserFeed_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/Search/users": {
      "get": {
        "tags": [
          "Search"
        ],
        "operationId": "Search_SearchUsers",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Search_SearchUsers_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Search_SearchUsers_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Search_SearchUsers_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Search_SearchUsers_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/Users/search": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Search",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_Search_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_Search_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_Search_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_Search_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetById_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetById_404ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetById_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetById_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetById_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetById_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/Users/{id}/followers": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetFollowers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProfileResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetFollowers_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetFollowers_400ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetFollowers_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "$ref": "#/components/schemas/Users_GetFollowers_401ErrorCode"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionResponse": {
        "required": [
          "id",
          "name",
          "description",
          "isPublic",
          "isSystem",
          "itemsCount",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "isSystem": {
            "type": "boolean"
          },
          "itemsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Collections_GetUserCollections_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Collections_GetUserCollections_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
...
      "Feed_GetUserFeed_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Feed_GetUserFeed_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
...
      "Search_SearchUsers_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Search_SearchUsers_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
...
      "UserProfileResponse": {
        "required": [
          "id",
          "displayName",
          "bio",
          "avatarUrl",
          "followersCount",
          "followingCount",
          "publicationsCount",
          "isFollowing"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "bio": {
            "type": [
              "null",
              "string"
            ]
          },
          "avatarUrl": {
            "type": "string"
          },
          "followersCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "followingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "publicationsCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": "integer",
            "format": "int32"
          },
          "isFollowing": {
            "type": "boolean"
          }
        }
      },
...
      "Users_GetById_404ErrorCode": {
        "enum": [
          "User.NotFound"
        ],
        "type": "string"
      },
...
      "Users_Search_400ErrorCode": {
        "enum": [
          "General.Validation"
        ],
        "type": "string"
      },
      "Users_Search_401ErrorCode": {
        "enum": [
          "Auth.Unauthorized"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}